Fixed illegal string operation error on video page (see desc)
This error was occurring on videos that did no have a thumbnail because the template was checking for the existence of a video url on a thumbnail that didn't exist. I updated the logic to check for a thumbnail only since the URL will exist inherently if the thumbnail exists.